refactor: do the ropm-modules check in a spec file
Replaces the standalone script + npm-script wiring with a single new spec
file. Mocha already loads every *.spec.ts, so the check runs at import
time and exits before any test can fail confusingly. Also point the
vscode build-test-project task at `build`, which runs `ropm copy`.
