+++ .npmignore
@@ -11,7 +11,6 @@ bsconfig.json
 .github
 .eslintrc.js
 test-project
-scripts
 docs
 apiDocs
 bsfmt.json
